requirements.txt
google-generativeai
scikit-learn
numpy
prompt_toolkit
flask
mysql